+++ libraries/ThreadAPI/include/thread/Timer.hpp
@@ -5,37 +5,30 @@
 
 #include <signal.h>
 
-#include "chrono/ClockTime.hpp"
 #include "Signal.hpp"
+#include "chrono/ClockTime.hpp"
 
 namespace thread {
 
 class Timer : public api::ExecutionContext {
 public:
   using ClockId = chrono::ClockTime::ClockId;
-  Timer(Signal::Event& signal_event, ClockId clock_id = ClockId::realtime);
+  Timer(Signal::Event &signal_event, ClockId clock_id = ClockId::realtime);
   ~Timer();
 
-  Timer(const Timer & a) = delete;
-  Timer& operator=(const Timer &a) = delete;
+  Timer(const Timer &a) = delete;
+  Timer &operator=(const Timer &a) = delete;
 
-  Timer(Timer && a){
-    std::swap(m_timer, a.m_timer);
-  }
+  Timer(Timer &&a) { std::swap(m_timer, a.m_timer); }
 
-  Timer& operator=(Timer &&a){
+  Timer &operator=(Timer &&a) {
     std::swap(m_timer, a.m_timer);
     return *this;
   }
 
-  bool is_valid() const {
-    return m_timer != timer_t(-1);
-  }
+  bool is_valid() const { return m_timer != timer_t(-1); }
 
-  enum class Flags {
-    null = 0,
-    absolute_time = TIMER_ABSTIME
-  };
+  enum class Flags { null = 0, absolute_time = TIMER_ABSTIME };
 
   class SetTime {
     API_AC(SetTime, chrono::ClockTime, interval);
@@ -43,7 +36,7 @@ public:
     API_AF(SetTime, Flags, flags, Flags::null);
   };
 
-  Timer & set_time(const SetTime & options);
+  Timer &set_time(const SetTime &options);
 
   class Info {
     API_RAC(Info, chrono::ClockTime, interval);
@@ -53,9 +46,21 @@ public:
 
   Info get_info() const;
 
+  class Alarm {
+  public:
+    enum class Type { seconds, useconds };
+
+  private:
+    API_AC(Alarm, chrono::ClockTime, value);
+    API_AC(Alarm, chrono::ClockTime, interval);
+    API_AF(Alarm, Type, type, Type::seconds);
+  };
+
+  static chrono::ClockTime alarm(const Alarm &options);
+  static void cancel_alarm() { alarm(Alarm()); }
+
 private:
   timer_t m_timer = timer_t(-1);
-
 };
 
 } // namespace thread