+++ libraries/ThreadAPI/src/Timer.cpp
@@ -1,4 +1,8 @@
+#include <unistd.h>
+
 #include "thread/Timer.hpp"
+
+extern "C" useconds_t ualarm(useconds_t useconds, useconds_t interval);
 
 using namespace thread;
 
@@ -21,10 +25,10 @@ Timer &Timer::set_time(const SetTime &options) {
   struct itimerspec value = {.it_interval = *options.interval().timespec(),
                              .it_value = *options.value().timespec()};
 
-  API_SYSTEM_CALL("", timer_settime(m_timer, int(options.flags()), &value, nullptr));
+  API_SYSTEM_CALL(
+      "", timer_settime(m_timer, int(options.flags()), &value, nullptr));
   return *this;
 }
-
 
 Timer::Info Timer::get_info() const {
   API_RETURN_VALUE_IF_ERROR(Info());
@@ -37,3 +41,14 @@ Timer::Info Timer::get_info() const {
   return result;
 }
 
+chrono::ClockTime Timer::alarm(const Alarm &options) {
+  API_RETURN_VALUE_IF_ERROR(chrono::ClockTime());
+  if (options.type() == Alarm::Type::seconds) {
+    API_SYSTEM_CALL("", ::alarm(options.value().seconds()));
+    return chrono::ClockTime(return_value() * 1_seconds);
+  }
+
+  API_SYSTEM_CALL("", ::ualarm(options.value().nanoseconds() / 1000UL,
+                               options.interval().nanoseconds() / 1000UL));
+  return chrono::ClockTime(return_value() * 1_microseconds);
+}